finished slide 14 challenge

+++ Ang_Raina-GDENG2/Ang_Raina-GDENG2/CubeObject.h
@@ -6,7 +6,7 @@ public:
 	CubeObject(std::string name);
 
 	void Initialize(void* shaderByteCode, size_t sizeShader, Vector3D color) override;
-	void Update(float deltaTime) override;
+	void Update(double deltaTime) override;
 
 	~CubeObject();
 };
